fix atoi

+++ S19/Core3/push_swap/libftplus/src/converter/ft_atoi.c
@@ -12,11 +12,11 @@
 
 #include "../../include/libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	char	sign;
 	int		i;
-	int		res;
+	long	res;
 
 	i = 0;
 	res = 0;
